refactoring indentation, also change accountId from int to long in all API Project(s)

+++ src/Projects/CreateProject.cs
@@ -2,7 +2,9 @@
 
 public partial class BasecampApiClient
 {
-    public async Task<Error?> CreateProjectAsync(int accountId, CreateProject data,
+    public async Task<Error?> CreateProjectAsync(
+        long accountId,
+        CreateProject data,
         CancellationToken cancellationToken = default)
     {
         if (!TokenHasBeenSet)
